# go_api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for the PET application",
        "title": "PET API",
        "contact": {
            "name": "API support",
            "email": "edsonjaybcarreon@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/",
    "paths": {
...
        "/persons/:id": {
            "get": {
                "description": "Get user by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "person"
                ],
                "summary": "Get a user by its ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Person ID",
                        "name": "person_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
...
}